WIP cleaning up relationships and controllers

+++ src/main/java/com/example/rentacar/services/implementations/CarServiceImplementation.java
@@ -3,8 +3,11 @@ package com.example.rentacar.services.implementations;
 import com.example.rentacar.DTOs.CarDTO;
 import com.example.rentacar.domain.Car;
 import com.example.rentacar.domain.CarType;
+import com.example.rentacar.domain.Rent;
+import com.example.rentacar.exceptions.BadRequestException;
 import com.example.rentacar.exceptions.CurrentlyRentedCarException;
 import com.example.rentacar.repositories.CarTypeRepository;
+import com.example.rentacar.repositories.RentRepository;
 import com.example.rentacar.services.ICarService;
 import com.example.rentacar.repositories.CarRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -12,6 +15,8 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.management.BadAttributeValueExpException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,9 +26,12 @@ public class CarServiceImplementation implements ICarService {
     @Autowired
     CarRepository carRepository;
     @Autowired
+    RentRepository rentRepository;
+    @Autowired
     ModelMapper modelMapper;
     @Autowired
     CarTypeRepository carTypeRepository;
+
 
     @Override
     public List<Car> getAlLCars() {
@@ -86,18 +94,39 @@ public class CarServiceImplementation implements ICarService {
 
     }
 
+
+    public boolean checkIfCurrentlyRented(List<Rent> rents) {
+        LocalDateTime currentDateTime = LocalDateTime.now();
+        for (Rent rent:rents) {
+            if(currentDateTime.isAfter(rent.getStartRent())&& currentDateTime.isBefore(rent.getEndRent()) )
+            {   return true;
+
+            }
+        }
+        return false;
+
+    }
+
+
     @Override
-    public void deleteCar (long carId) {
+    public void deleteCar (long carId)  {
 
+            boolean isCurrentlyRented = false;
+            Car carToDelete = carRepository.findById(carId).orElse(null);
+            if(carToDelete==null){
+               throw new BadRequestException("No car has been found with the ID provided");
+            }
 
-        try {
-            carRepository.deleteById(carId);
+            if (carToDelete != null) {
+                List<Rent> rents = carToDelete.getRentList();
+                isCurrentlyRented = checkIfCurrentlyRented(rents);
+                if (isCurrentlyRented) {
+                    throw new CurrentlyRentedCarException("The car is currently under rental and cannot be deleted from the database.");
+                } else {
+                    carRepository.deleteById(carId);
 
-        }
-        catch(Exception e) {
-
-            throw new CurrentlyRentedCarException(e.getMessage());
-        }
+                }
+            }
 
 
     }
